Change the way cells will be rendered
We used to do:

         [{
          "activity": {
            "code": "cospro_c_5_1",
            "name": "COSPRO - C.5.1 - 100% des écoles  des réseaux sont controlés sur leur gestion financière"
          },
          "cells": {
            "available_budget": {
              "key": "cospro__c_5__gestion_financiere_cospro_c_5_1_available_budget_for_AOsKyLAjVWH_and_2019q1",
              "instantiated_expression": "0",
              "not_exported": false,
              "solution": "0",
              "substituted": "0",
              "is_input": true,
              "is_output": false,
              "state": {
                "ext_id": "rRwPsKFXTX6",
                "kind": "data_element",
                "name": "COSPRO - Budget Disponible - C.5.1 - 100% des écoles  des réseaux sont controlés sur leur gestion financière"
              }
            }]

This will change the `activity` key to be `meta` and it will change
the cells, to be up one level. So instead of `cells ->
available_budget`, it will be `available_budget` and then the facts.

I've also made every item in that list have a `key` and a `value`, so
when looping over them you can always assume those two are there. I've
kept `solution` for now.