Expose a sets endpoints
This will return all info about a set that we currently have (if there
is anything missing it will be added but the structure seems to be
this)

It's a JSON-API compound document, so you have your normal:

data: [
  { id: "1",
    type: "set"
    attributes: {
      // List of attrubites on the set
    }
  },
  "relationships": {
    // Currently listing: org_unit_groups, org_unit_group_sets,
    //  inputs, topics and topic_formulas
  }
]

Now what makes it a compound document, is that the info for the
relationships is included, next to the data leaf is an included leaf
which contains all the data needed to render the relationships:

included: [
  {
      "id": "1",
      "type": "input",
      "attributes": {
        "name": "Claimed",
        "short_name": null
      }
    },
]

So if you need to find the info about input with id 1, you need to
look in the `included` array and do a search for `id=1` and
`type=input`, and then you can retrieve the attributes without having
to make an additional call.

If you use:
https://github.com/itsfadnis/jsonapi-serializer#deserialization, this
will do that for free for you.

I tested this with:

      const JSONAPIDeserializer = require('jsonapi-serializer').Deserializer;
      const sets = require("./sets.json")
      new JSONAPIDeserializer().deserialize(sets, (e, u) => console.log(u));
      // THis will show the entire output, but will show [object object] for
      // the relationships
      new JSONAPIDeserializer().deserialize(sets, (e, u) => console.log(u[0].inputs));

      [ { name: 'Claimed', 'short-name': null, id: '1' },
        { name: 'Max. Score', 'short-name': null, id: '4' },
        { name: 'Budget', 'short-name': null, id: '6' } ]

So it already does the lookups, and you don't need to do it anymore.