refactor: More production like data for mock server (#52)
Our mocked data was based on various factories in our test data, this lead to some discrepancies between the different models. Now we're using a new approach where I took a snapshot of an actual project, and then carefully switched some of the data so that we don't need an external S3 to host the blobs.